refactor: reverse the event manager implementation split
